Update to file structure
Added components folder and newsletter.js file under it then added the newsletter component to the components page
