--- src/pages/components/newsletter.js
@@ -1,5 +1,0 @@
-export default function Components() {
-    return (
-      <h1>components</h1>
-    );
-  }
